Simplify service worker
It always hurts to give up something you worked hard to get running, but
I don't think that this elaborate service worker has a place in this app.
The incremental updating where it only re-downloads files is – in theory –
really nice for bandwith, but this morning I took the size of the
minified app, and it came out to about 30kB (without the service worker,
which was about 10kB minified). This is just not worth all this effort
and complexity. Unfortunately, downsizing the service worker did not
improve load times :/  I'm afraid that the long load is just inherent
in the technology, which is annoying.