improve hash table format for access speed
previously, the hash table used open adressing to handle collisions,
which meant that in some cases, it had to iterate over quite a few
entries (since with open adressing, entries from different buckets can
collide and form big chunks of entries). since the keys are not
guaranteed to be unique, we had to search the entire chunk (until
finding an empty bucket) to make sure we'd seen all entries. (this also
means that my implementation of robin hood hashing was basically
useless.)

the new format stores data in basically the same way, except that each
bucket now has a separate array of entries. (the entries are still
stored back-to-back, but now there is an association to the bucket in
which they belonged.) this means that we *only* have to search through
collisions, and not through an entire chunk of entries from a number of
buckets, which should be a lot fewer. (especially due to the high load
factor.) this does lead to an increase in memory usage, since each
definition now needs an extra 2 bytes to store its length, so we can
quickly jump to the next one if the current one didn't match. (maybee
that's not that important though?) we didn't need this previously, since
all entries were referenced directly from the hash table.

unfortunately, it's apparently really really hard to get reliable
performance info, since it has *huge* variance. i'm not really sure why
this is or how to deal with it :/