Add aria-labelledby
Since the entry box is implicitly labelled by the mode header,
I thought it might be a good idea to use aria-labelledby to
formalize this relationship. I am a bit worried that this
solution might be too fancy and that I should really just
add a normal label to the text box, but I'm not quite sure
how to make that look good and maybe I'm worrying too much
about this.