move to workspaces so we can test hashtable separately
things are kind of annoying to test with no_std and a panic handler set, plus
the main module relies on wasm intrinsics. specifically, the main module
sets a panic_handler, which interferes with normal testing, which includes
std, which has its own conflicting panic handler. in principle, the hashtable
module is independent of this, but without putting it in a separate crate
it'll still get compiled along with the panic handler. we solve this by
using a workspace and moving hashtable to a separate package.