fix the performance bug
this missing line basically meant that all keys got hashed into the same bucket
(i think that because of the stroke formatting maybe the last byte was always zero
or something, in addition to this problem?). this in turn drastically increased
runtime as all 150000 entries were all inserted into a single bucket.