fix bug in ParseStrokesIterator
previously, ParseStrokesIterator would shift out bytes
one by one and start emitting the new stroke once
once the internal state was 0. this meant that if the
third byte of a stroke was zero (the fourth isn't used
and is always zero), it would only emit the first two,
which is kind of a problem for reading the strokes
back out.